workqueue: Split out workqueue_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performace

More sched.h dependency culling - this lets us kill a rhashtable-types.h
dependency on workqueue.h.

Conflicts:
	include/linux/workqueue.h
(cherry picked from commit b2fa8443db320c4873feca2588b957439e350890)
